openai-agent/pyproject.toml
[project]
name = "openai-agent"
version = "0.1.0"
description = "OpenAI Agent SDK - The Minimalist for GPT-native performance"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


